third_party/WebKit/LayoutTests/media/W3C/audio/canPlayType/canPlayType_application_octet_stream_with_codecs_2-expected.txt

audio.canPlayType() - can never play application/octet-stream with a codecs parameter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is ""
PASS successfullyParsed is true

TEST COMPLETE
spec reference
